refactor screen file structure
