+++ src/screens/auth/LoginScreen.tsx
@@ -1,12 +1,12 @@
 import React, { useState } from 'react';
-import { Button, Snackbar, TextInput } from '../components';
-import { SUPABASE } from '../helpers';
-import { useSessionContext } from '../hooks';
+import { Button, Snackbar, TextInput } from '../../components';
+import { SUPABASE } from '../../helpers';
+import { useSessionContext } from '../../hooks';
 import { Link, Navigate } from 'react-router-dom';
 import { InputAdornment, IconButton, Typography as Typ } from '@mui/material';
 import { Visibility, VisibilityOff } from '@mui/icons-material';
-import Logger from '../logger';
-import { SnackbarProps } from '../components/Snackbar';
+import Logger from '../../logger';
+import { SnackbarProps } from '../../components/Snackbar';
 
 const LOG = new Logger('LoginForm');
