+++ src/screens/auth/index.ts
@@ -1,0 +1,5 @@
+import AuthLayout from './AuthLayout';
+import LoginScreen from './LoginScreen';
+import SignUpScreen from './SignUpScreen';
+
+export { AuthLayout, LoginScreen, SignUpScreen };
