+++ src/screens/dashboard/index.ts
@@ -1,0 +1,5 @@
+import DashboardLayout from './DashboardLayout';
+import DashboardScreen from './DashboardScreen';
+import DashboardGalleryScreen from './DashboardGalleryScreen';
+
+export { DashboardLayout, DashboardScreen, DashboardGalleryScreen };
